+++ metrics/init.lua
@@ -96,6 +96,9 @@ return {
         require('metrics.tarantool.luajit').enable()
         require('metrics.default_metrics.tarantool').enable()
     end,
+    enable_cartridge_metrics = function()
+        return require('metrics.cartridge_metrics').enable()
+    end,
     http_middleware = require('metrics.http_middleware'),
     collect = collect,
 }